Make collectionIndexing aa util
